Fixed: Don't reduce permeability below base material
In cracks, permeability is modeled according to Poiseuille flow, however it's
entirely possible for the crack to be narrow enough so that Poiseuille flow
gives a result that is less permeable than the base material. In those cases,
the fluid should be allowed to flow through the pores rather than the crack
at unmodified rate.